Create evolution.txt
An attempt to devise format for storing evolution path for habits.
There are many things-awesome appear every day, but it there
is no database that can be used to lookup and upgrade your
habits automatically.

Feel free to evolve and create your own format. The format is
detected by the first line, but it can involve too.

    # evolution: techtonik.1

In the end what matters is ability to evolve, not standards.
But I put this right, tools can help to evolve from techtonik.1
to next yourformat without problem.